+++ dataprovider/basketballreference/nba/shared.go
@@ -23,7 +23,7 @@ var networkHeaders network.Headers = network.Headers(map[string]interface{}{
 	"user-agent":                "Mozilla/5.0 (Macintosh; Intel Mac OS X 10_15_7) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/103.0.5060.53 Safari/537.36",
 })
 
-type headerValues map[string]struct{}
+type headerValues []string
 
 const (
 	// https://www.basketball-reference.com/boxscores/{event_id}.html